limit processing to events we actually listen for
by default, the webhook handler from django-github-app processes and stores all events, even if they are not registered as actionable.

since the Pull Request event on GitHub sends many events we are not interested in:

> Pull request assigned, auto merge disabled, auto merge enabled, closed, converted to draft, demilestoned, dequeued, edited, enqueued, labeled, locked, milestoned, opened, ready for review, reopened, review request removed, review requested, synchronized, unassigned, unlabeled, or unlocked.

This causes floods of events in situations like https://github.com/python/cpython/pull/132964 that we can skip processing for.